requirements.txt
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
chispa==0.10.0
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
delta-spark==3.1.0
executing==2.0.1
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pyspark==3.5.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
setuptools==69.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
zipp==3.18.1
